💄 design: 아티스트 바텀시트 pc 모달 변환 적용
